requirements.txt
--find-links https://storage.googleapis.com/jax-releases/jax_cuda_releases.html

mujoco==2.3.7
numpy==1.26.4
scipy==1.13
jax[cuda12_pip]==0.4.26
flax==0.8.4
git+https://github.com/JoeyTeng/brax.git@main#egg=brax
scikit-learn==1.5.2
statsmodels==0.14.2
seaborn==0.13.2
ipython==8.27.0
ipykernel==6.29.5
imageio==2.35.1
hydra-core==1.3.2
